+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 All Notable changes to `Manager` will be documented in this file
 
-# NEXT 
- - Initialize `CollectsItemsTrait` and `Collection` class
- - Initialize `IocContainerTrait` for basic DI functionality
+# NEXT (0.8.5)
+ - Initialize `CollectsItemsTrait` and `Collection` class as experimental
+ - Initialize `IocContainerTrait` for basic DI functionality as experimental
 
 #v0.8.4 - 2015-7-14
  - Better documentation (README.md)